fishing-game/server/requirements.txt
Flask>=2.0
Flask-SocketIO>=5.0
python-dotenv>=0.19
pymongo>=4.0
pydantic>=2.0 # BaseSettings moved out in v2
pydantic-settings>=2.0 # For BaseSettings
# For eventlet or gevent (needed by Flask-SocketIO for production/async)
# Choose one based on preference/needs later
# eventlet
# gevent
# gevent-websocket
